Fix a merge-hez

+++ framework/logger/DatabaseWriter.cpp
@@ -6,7 +6,7 @@ void DatabaseWriter::_writeLog(const cSeverity::teSeverity sev, const QDateTime 
     {
         try {
             QString  ssSql;
-            ssSql = QString("INSERT INTO logs (userId, severity, message, date, threadId) VALUES('%1', '%4', '%2', '%3', '%5');").arg(m_uiAppUser).arg(m).arg(ts.toString("yyyy-MM-dd hh:mm:ss")).arg(sev).arg(reinterpret_cast<int>(threadId), 0, 16);
+            ssSql = QString("INSERT INTO logs (userId, severity, message, date, threadId) VALUES('%1', '%4', '%2', '%3', '%5');").arg(m_uiAppUser).arg(m).arg(ts.toString("yyyy-MM-dd hh:mm:ss")).arg(sev).arg(reinterpret_cast<long>(threadId), 0, 16);
             m_poDB->executeQTQuery( ssSql );
         } catch(cSevException &e)
         {
